aangepast

+++ DBLaag/Product.cs
@@ -8,7 +8,7 @@ namespace DBLaag
     {
         public int Id { get; set; }
         public string Naam { get; set; }
-        public int Prijs { get; set; }
+        public double Prijs { get; set; }
         public string Specificaties { get; set; }
         public string Fotonaam { get; set; }
         public int Discount { get; set; }
